copia diaria

+++ employeesService/src/main/java/com/init/employees/dao/IEmployeeDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.init.employees.dto.Employee;
 
-public interface IEmployeeDAO extends JpaRepository<Employee, Long>{
+public interface IEmployeeDAO extends JpaRepository<Employee, Integer>{
 
 }
