+++ employeesService/src/main/java/com/init/employees/dao/IJobDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.init.employees.dto.Job;
 
-public interface IJobDAO extends JpaRepository<Job, Long>{
+public interface IJobDAO extends JpaRepository<Job, Integer>{
 
 }
