+++ employeesService/src/main/java/com/init/employees/service/IJobService.java
@@ -9,8 +9,8 @@ public interface IJobService {
 	//Metodos CRUD
 	public List<Job> listJob(); //Llistar All
 	public Job saveJob(Job job); //Guarda Job
-	public Job jobXID(Long id); //Leer datos de un trabajo por id
+	public Job jobXID(Integer id); //Leer datos de un trabajo por id
 	public Job updateJob(Job job); //Actualizar job
-	public void deleteJob(Long id); //elimina job
+	public void deleteJob(Integer id); //elimina job
 
 }
